Start using tailwind.css
Tailwind is a much easier way to structure this site for different form
factors such as phones and small laptops.

Also add a button to scroll to the about section. Doing so requires
a modern browser which offers smooth scrolling.

Thanks to @tanish2k09 for helping out with reconfiguring the site.

Signed-off-by: kingbri <bdashore3@gmail.com>