Add a navigation bar
A horizontal navigation bar is best for this site. The bar is sticky
from the "About Me" page all the way down.

On hover, a neat underlining animation pops up which is not too glaring,
but also clean.

I was inspired by [this codepen](https://codepen.io/jstn/pen/zuDst) for
the underlining effect.

Also, make all pages use a width of screen due to div sizing issues.

Signed-off-by: kingbri <bdashore3@gmail.com>