Add socials footer
Instead of having a full-fledged contact me page, a socials footer
is much easier to work with.

This footer is designed to be small, but not forgotten. People
can navigate to my links by hovering over the SVG images. The
reason SVG images are used is because it's easier to work with on
webpages.

As a result, extra images/routes are removed from the main.rs file.

Signed-off-by: kingbri <bdashore3@gmail.com>