Affix navbar to the top of the page
Fixing the navbar to the top makes it easier for people to navigate
the website without pushing the arrow button on the bottom. This
is useful if people want to navigate to a submenu instead of
scrolling down to do so.

In addition, take some time to fix the homepage's layout to something
simpler with flexboxes

Signed-off-by: kingbri <bdashore3@gmail.com>