Components: Split redundancies into classes
The original HTML of this website contained a lot of redundant code.
Now that I'm using Next.js, split the base elements into a function
and use that when rendering specific elements on the page.

Use AnchorLink for on-page links such as `#about` and `#work` to keep
the URL clean. The URL should only contain anything extra if it's explicitly
stated.

Signed-off-by: kingbri <bdashore3@gmail.com>