Blog: Fix post updating mechanism
When fallback isn't enabled on the blog, the posts do not regenerate
through SSG since there is no stimulus to reload the post content.

This results on a post entry showing up on the blog homepage, but
a 404 is fired when the user clicks on the post since the content wasn't
populated at build time.

Signed-off-by: kingbri <bdashore3@gmail.com>